requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alpaca-trade-api==3.2.0
altair==5.5.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
curl_cffi==0.10.0
cycler==0.12.1
deprecation==2.1.0
exceptiongroup==1.2.1
executing==2.0.1
fonttools==4.58.0
frozendict==2.4.6
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
msgpack==1.0.3
multidict==6.4.3
multitasking==0.0.11
narwhals==1.38.2
nest_asyncio==1.6.0
numpy==2.2.5
packaging==24.2
pandas==2.2.3
peewee==3.18.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
platformdirs==4.2.1
prompt_toolkit==3.0.43
propcache==0.3.1
protobuf==6.30.2
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.2
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.2.3
python-dateutil==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.16.0
smmap==5.0.2
soupsieve==2.7
stack_data==0.6.2
streamlit==1.45.0
ta==0.11.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
websockets==10.4
yarl
yfinance==0.2.58
zipp==3.18.1